+++ frontend/src/components/classroom/Classroom.jsx
@@ -10,10 +10,8 @@ import { BiEdit } from "react-icons/bi";
 import { Form, useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
 import { elements } from "../../enums/Element";
-import { tipologies } from "../../enums/Tipology";
 import { getClassrooms, deleteClassroomById } from "../../api/ClassroomApiService";
 import Swal from "sweetalert2";
-
 
 
 
@@ -21,36 +19,21 @@ import Swal from "sweetalert2";
 /**
  * La función reformatElement toma el valor de un elemento y devuelve su nombre correspondiente de un
  * lista predefinida de elementos.
- * @returns La función `reformatElement` toma un argumento `element` y busca un objeto en un
- * matriz llamada `elementos` que tiene una propiedad `valor` igual al argumento `elemento`. Si tal
- * se encuentra el objeto, la función devuelve el valor de su propiedad `nombre`. De lo contrario, una cadena vacía
- * es regresado.
+ * @returns La función `reformatElement` toma un argumento `element` y busca un objeto en
+ * la matriz `elementos` que tiene una propiedad `valor` igual a `elemento`. Si se encuentra tal objeto, el
+ * la función devuelve el valor de su propiedad `name`. Si no, se devuelve una cadena vacía.
  */
 const reformatElement=(element)=>{
 
   const foundElement = elements.find((e) => e.value === element);
   return foundElement ? foundElement.name : "";
- }
-
-
-/**
-  * La función reformatTipology toma un valor de tipología y devuelve su nombre correspondiente de un
-  * lista predefinida de tipologías.
-  * @returns La función `reformatTipology` toma un parámetro `tipología` y devuelve el correspondiente
-  * Propiedad `nombre` de la matriz `tipologías` si se encuentra una propiedad `valor` coincidente. Si no hay coincidencia
-  * encontrado, se devuelve una cadena vacía.
-  */
- const reformatTipology=(tipology)=>{
-
-  const foundTipology = tipologies.find((e) => e.value === tipology);
-  return foundTipology ? foundTipology.name : "";
   }
-  
+
+
 
 
 /**
- * El código anterior define dos funciones para mostrar alertas de éxito y error utilizando la biblioteca Swal
- * en JavaScript Reaccionar.
+ * La función muestra un mensaje de alerta de éxito con el nombre de un salón eliminado.
  */
   const succesResponseAlert= (response)=>{
     Swal.fire(
@@ -59,6 +42,11 @@ const reformatElement=(element)=>{
       'success',
     )
   }
+
+
+/**
+  * La función muestra una alerta de error con un mensaje de la respuesta de error.
+  */
   const errorResponseAlert = (error)=>{
     Swal.fire({
       title: "Error",
@@ -71,42 +59,41 @@ const reformatElement=(element)=>{
   }
 
 export const Classroom = () => {
-
-/* El código define tres variables de estado usando el gancho `useState`: `classroom`, `currentPage`,
- y `aulasCambio`. */
+  
+/* Estas líneas de código definen e inicializan variables de estado utilizando el gancho `useState`. */
   const [classroom, setClassroom] = useState([]);
-
   const [currentPage, setCurrentPage] = useState(1);
   const [classroomsChange, setClassroomsChange] = useState(false);
   const classroomsPerPage = 5;
+
+  const [search, setSearch] = useState("");
+/**
+   * La función establece los datos del aula a partir de una respuesta exitosa.
+   */
   const succesResponses = (res) => {
     setClassroom(res.data);
   };
+
+  /*`useEffect` se usa para obtener la lista de aulas de la API usando el
+  función `getClassrooms` y actualice el estado de la variable `classroom` con los datos de respuesta
+  utilizando la función `setClassroom`. La función `succesResponses` se llama cuando la llamada API es
+  exitoso y actualiza el estado de la variable `aula`. Si hay un error, el `error`
+  Se llama a la función y registra el error en la consola. La variable `classroomsChange` se usa como
+  una dependencia para el gancho `useEffect`, lo que significa que el efecto se volverá a ejecutar siempre que el
+  cambia el valor de `classroomsChange`. La línea `setClassroomsChange(false)` se utiliza para restablecer el
+  valor de `classroomsChange` a `false` después de que se haya ejecutado el efecto. */
   useEffect(() => {
     getClassrooms()
       .then((response) => succesResponses(response))
       .catch((error) => console.error(error));
       setClassroomsChange(false);
   }, [classroomsChange]);
-
-
-  //-------------------------------------------------------------------------
-
   const lastClassroomIndex = currentPage * classroomsPerPage;
   const firstClassroomIndex = lastClassroomIndex - classroomsPerPage;
   let currentClassroom = classroom.slice(firstClassroomIndex, lastClassroomIndex);
 
-/* El código define una función llamada `deleteClassroomHandler` que toma un parámetro `id`. Este
-La función se usa para manejar la eliminación de un salón de clases llamando a la función `deleteClassroomById`
-del módulo `ClassroomApiService`. Antes de eliminar el aula, aparece un cuadro de diálogo de confirmación.
-se muestra usando la función `Swal.fire` de la biblioteca SweetAlert2. Si el usuario confirma la
-eliminación, se llama a la función `deleteClassroomById` y si la eliminación es exitosa, un éxito
-La alerta se muestra usando la función `succesResponseAlert`. Si hay un error durante la
-eliminación, se muestra una alerta de error utilizando la función `errorResponseAlert`. Finalmente, el
-Se llama a la función `setClassroomsChange` para actualizar el estado de la variable `classroomsChange`,
-que desencadena una nueva representación del componente. */
   const navigate = useNavigate();
-  const deleteClassroomHandler = (id) => {
+  const deleteSubjectHandler = (id) => {
     Swal.fire({
       title: '¿Estas seguro de eliminar este salon?',
       text: "Estos cambios son irreversibles",
@@ -122,7 +109,7 @@ que desencadena una nueva representación del componente. */
           deleteClassroomById(id)
             .then((response) => {
               succesResponseAlert(response);
-              setClassroomsChange(true)
+              setClassroomsChange(true) //Indicates to the useEffect to update the subject list
             }
             )
             .catch((error) => {
@@ -132,16 +119,9 @@ que desencadena una nueva representación del componente. */
       })
 
     }
-    //-------------------------------------------------------------------------
    
-
-/* Este código define una funcionalidad de búsqueda para el componente Classroom. Crea una variable de estado.
-  `buscar` y una función `searchHandler` que actualiza el estado de `búsqueda` con el valor del
-  campo de entrada. Luego, comprueba si la variable `buscar` no es una cadena vacía y filtra la
-  matriz `currentClassroom` basada en si la propiedad `name` de cada aula incluye el
-  Cadena de `búsqueda` (no distingue entre mayúsculas y minúsculas). Esto permite al usuario buscar un aula específica por
-  nombre. */
-    const [search, setSearch] = useState("");
+  
+  
     const searchHandler = (event)=>{
       setSearch(event.target.value);
     }
@@ -149,6 +129,8 @@ que desencadena una nueva representación del componente. */
     if(search.trim() !== ""){
       currentClassroom = classroom.filter(classroom=> classroom.name.toLowerCase().includes(search))
     }
+
+  
 
   return (
     <Flex
@@ -178,7 +160,7 @@ que desencadena una nueva representación del componente. */
             </Button>
           </div>
           {<Input
-            input={{ placeholder: "Nombre de la persona" ,onChange:searchHandler}}
+            input={{ placeholder: "Nombre del salon" ,onChange:searchHandler }}
             style={{ height: "20px" }}
           ></Input> }
           
@@ -190,7 +172,7 @@ que desencadena una nueva representación del componente. */
           </Button> }
        
         </Flex>
-        {currentClassroom.length > 0 ?
+        {classroom.length > 0 ?
           <>
         <table className={style.table}>
           <thead>
@@ -206,39 +188,30 @@ que desencadena una nueva representación del componente. */
           </thead>
           <tbody>
             {currentClassroom.map((classroom) => (
-               
               <tr key={classroom.id}>
                 <td className={style.id}>{classroom.id}</td>
                 <td>{classroom.name}</td>
                 <td>{classroom.location}</td>
                 <td>{classroom.capability}</td>
-               
                 <td style={{textAlign:"left"}}>
                   <ul>
                     {classroom.elements.map((element,index) => (
                       <li style={{fontSize:"20px"}} key={index}>{reformatElement(element)}</li>
-                      
                     ))}
+
                   </ul>
                 </td>
-                <td style={{textAlign:"left"}}>
-                  <ul>
-                 
-                  <li>{reformatTipology(classroom.tipology)}</li>
-               
-                  </ul> 
-                </td>
+                <td>{classroom.tipology}</td>
 
                 <td className={style["actions__container"]}>
                     <BiEdit className={style["icon__edit"]} onClick={() => navigate(`/salones/editar/${classroom.id}`)} />
-                    <MdDeleteForever className={style["icon__delete"]}  onClick={deleteClassroomHandler.bind(null, classroom.id)}/>
+                    <MdDeleteForever className={style["icon__delete"]}  onClick={deleteSubjectHandler.bind(null, classroom.id)}/>
                 </td>
               </tr>
             ))}
 
           </tbody>
         </table>
-      
         {classroom.length > 8 && (
           <Pagination
             totalItems={classroom.length}
